fix: current projects shows data
I had to change the API uri by removing the sheet name and adding A1:Y38 (the range of values we want in the table). I randomly guessed this and it worked.

I believe it is because we only have one sheet in the spreadsheet so we only have to specify the range of values we want to get.